re-record tortoise metrics when phase or mode is changed
