+++ pkg/metrics/metrics.go
@@ -3,8 +3,6 @@ package metrics
 import (
 	"github.com/prometheus/client_golang/prometheus"
 	"sigs.k8s.io/controller-runtime/pkg/metrics"
-
-	"github.com/mercari/tortoise/api/v1beta3"
 )
 
 var (
@@ -79,19 +77,4 @@ func init() {
 		ProposedMemoryRequest,
 		TortoiseNumber,
 	)
-}
-
-func RecordTortoise(t *v1beta3.Tortoise, deleted bool) {
-	value := 1.0
-	if deleted {
-		value = 0
-	}
-	TortoiseNumber.WithLabelValues(
-		t.Name,
-		t.Namespace,
-		t.Spec.TargetRefs.ScaleTargetRef.Name,
-		t.Spec.TargetRefs.ScaleTargetRef.Kind,
-		string(t.Spec.UpdateMode),
-		string(t.Status.TortoisePhase),
-	).Set(value)
 }
